core: move container exec code to own file
container.go is understandably a pretty big file but the exec related
code is the most complicated and interesting part, which makes it
annoying to always have to find amongst everything else. Hence, it feels
a bit nicer to have in its own dedicated file.
