+++ core/containerExec.go
@@ -1,0 +1,292 @@
+package core
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"os"
+	"sort"
+	"strings"
+
+	"github.com/dagger/dagger/dagql"
+	"github.com/dagger/dagger/engine"
+	"github.com/dagger/dagger/engine/buildkit"
+	"github.com/moby/buildkit/client/llb"
+	"github.com/pkg/errors"
+)
+
+func (container *Container) WithExec(ctx context.Context, opts ContainerExecOpts) (*Container, error) { //nolint:gocyclo
+	container = container.Clone()
+
+	cfg := container.Config
+	mounts := container.Mounts
+	platform := container.Platform
+	if platform.OS == "" {
+		platform = container.Query.Platform
+	}
+
+	args, err := container.command(opts)
+	if err != nil {
+		return nil, err
+	}
+
+	var namef string
+	if container.Focused {
+		namef = buildkit.FocusPrefix + "exec %s"
+	} else {
+		namef = "exec %s"
+	}
+
+	runOpts := []llb.RunOption{
+		llb.Args(args),
+		llb.WithCustomNamef(namef, strings.Join(args, " ")),
+	}
+
+	// this allows executed containers to communicate back to this API
+	var clientID string
+	var otelEnvs []string
+	if opts.ExperimentalPrivilegedNesting {
+		callerOpts := opts.NestedExecFunctionCall
+		if callerOpts == nil {
+			// default to caching the nested exec
+			callerOpts = &FunctionCall{
+				Cache: true,
+			}
+		}
+		clientID, err = container.Query.RegisterCaller(ctx, callerOpts)
+		if err != nil {
+			return nil, fmt.Errorf("register caller: %w", err)
+		}
+		otelEnvs = callerOpts.OTELEnvs
+
+		// include the engine version so that these execs get invalidated if the engine/API change
+		runOpts = append(runOpts, llb.AddEnv("_DAGGER_ENGINE_VERSION", engine.Version))
+
+		// include a digest of the current call so that we scope of the cache of the ExecOp to this call
+		runOpts = append(runOpts, llb.AddEnv("_DAGGER_CALL_DIGEST", string(dagql.CurrentID(ctx).Digest())))
+
+		if !callerOpts.Cache {
+			// include the ServerID here so that we bust cache once-per-session
+			clientMetadata, err := engine.ClientMetadataFromContext(ctx)
+			if err != nil {
+				return nil, err
+			}
+			runOpts = append(runOpts, llb.AddEnv("_DAGGER_SERVER_ID", clientMetadata.ServerID))
+		}
+	}
+
+	metaSt, metaSourcePath := metaMount(opts.Stdin)
+
+	// create /dagger mount point for the shim to write to
+	runOpts = append(runOpts,
+		llb.AddMount(buildkit.MetaMountDestPath, metaSt, llb.SourcePath(metaSourcePath)))
+
+	if opts.RedirectStdout != "" {
+		runOpts = append(runOpts, llb.AddEnv("_DAGGER_REDIRECT_STDOUT", opts.RedirectStdout))
+	}
+
+	if opts.RedirectStderr != "" {
+		runOpts = append(runOpts, llb.AddEnv("_DAGGER_REDIRECT_STDERR", opts.RedirectStderr))
+	}
+
+	for _, bnd := range container.Services {
+		for _, alias := range bnd.Aliases {
+			runOpts = append(runOpts,
+				llb.AddEnv("_DAGGER_HOSTNAME_ALIAS_"+alias, bnd.Hostname))
+		}
+	}
+
+	if cfg.User != "" {
+		runOpts = append(runOpts, llb.User(cfg.User))
+	}
+
+	if cfg.WorkingDir != "" {
+		runOpts = append(runOpts, llb.Dir(cfg.WorkingDir))
+	}
+
+	for _, env := range cfg.Env {
+		name, val, ok := strings.Cut(env, "=")
+		if !ok {
+			// it's OK to not be OK
+			// we'll just set an empty env
+			_ = ok
+		}
+
+		// don't pass these through to the container when manually set, they are internal only
+		if name == "_DAGGER_NESTED_CLIENT_ID" && !opts.ExperimentalPrivilegedNesting {
+			continue
+		}
+
+		runOpts = append(runOpts, llb.AddEnv(name, val))
+	}
+
+	// if GPU parameters are set for this container pass them over:
+	if len(container.EnabledGPUs) > 0 {
+		if gpuSupportEnabled := os.Getenv("_EXPERIMENTAL_DAGGER_GPU_SUPPORT"); gpuSupportEnabled == "" {
+			return nil, fmt.Errorf("GPU support is not enabled, set _EXPERIMENTAL_DAGGER_GPU_SUPPORT")
+		}
+		runOpts = append(runOpts, llb.AddEnv("_EXPERIMENTAL_DAGGER_GPU_PARAMS", strings.Join(container.EnabledGPUs, ",")))
+	}
+
+	secretsToScrub := SecretToScrubInfo{}
+	for i, secret := range container.Secrets {
+		secretOpts := []llb.SecretOption{llb.SecretID(secret.Secret.Accessor)}
+
+		var secretDest string
+		switch {
+		case secret.EnvName != "":
+			secretDest = secret.EnvName
+			secretOpts = append(secretOpts, llb.SecretAsEnv(true))
+			secretsToScrub.Envs = append(secretsToScrub.Envs, secret.EnvName)
+		case secret.MountPath != "":
+			secretDest = secret.MountPath
+			secretsToScrub.Files = append(secretsToScrub.Files, secret.MountPath)
+			if secret.Owner != nil {
+				secretOpts = append(secretOpts, llb.SecretFileOpt(
+					secret.Owner.UID,
+					secret.Owner.GID,
+					int(secret.Mode),
+				))
+			}
+		default:
+			return nil, fmt.Errorf("malformed secret config at index %d", i)
+		}
+
+		runOpts = append(runOpts, llb.AddSecret(secretDest, secretOpts...))
+	}
+
+	if len(secretsToScrub.Envs) != 0 || len(secretsToScrub.Files) != 0 {
+		// we sort to avoid non-deterministic order that would break caching
+		sort.Strings(secretsToScrub.Envs)
+		sort.Strings(secretsToScrub.Files)
+
+		secretsToScrubJSON, err := json.Marshal(secretsToScrub)
+		if err != nil {
+			return nil, fmt.Errorf("scrub secrets json: %w", err)
+		}
+		runOpts = append(runOpts, llb.AddEnv("_DAGGER_SCRUB_SECRETS", string(secretsToScrubJSON)))
+	}
+
+	for _, ctrSocket := range container.Sockets {
+		if ctrSocket.ContainerPath == "" {
+			return nil, fmt.Errorf("unsupported socket: only unix paths are implemented")
+		}
+
+		socketOpts := []llb.SSHOption{
+			llb.SSHID(ctrSocket.Source.SSHID()),
+			llb.SSHSocketTarget(ctrSocket.ContainerPath),
+		}
+
+		if ctrSocket.Owner != nil {
+			socketOpts = append(socketOpts,
+				llb.SSHSocketOpt(
+					ctrSocket.ContainerPath,
+					ctrSocket.Owner.UID,
+					ctrSocket.Owner.GID,
+					0o600, // preserve default
+				))
+		}
+
+		runOpts = append(runOpts, llb.AddSSHSocket(socketOpts...))
+	}
+
+	for _, mnt := range mounts {
+		srcSt, err := mnt.SourceState()
+		if err != nil {
+			return nil, fmt.Errorf("mount %s: %w", mnt.Target, err)
+		}
+
+		mountOpts := []llb.MountOption{}
+
+		if mnt.SourcePath != "" {
+			mountOpts = append(mountOpts, llb.SourcePath(mnt.SourcePath))
+		}
+
+		if mnt.CacheVolumeID != "" {
+			var sharingMode llb.CacheMountSharingMode
+			switch mnt.CacheSharingMode {
+			case CacheSharingModeShared:
+				sharingMode = llb.CacheMountShared
+			case CacheSharingModePrivate:
+				sharingMode = llb.CacheMountPrivate
+			case CacheSharingModeLocked:
+				sharingMode = llb.CacheMountLocked
+			default:
+				return nil, errors.Errorf("invalid cache mount sharing mode %q", mnt.CacheSharingMode)
+			}
+
+			mountOpts = append(mountOpts, llb.AsPersistentCacheDir(mnt.CacheVolumeID, sharingMode))
+		}
+
+		if mnt.Tmpfs {
+			mountOpts = append(mountOpts, llb.Tmpfs())
+		}
+
+		if mnt.Readonly {
+			mountOpts = append(mountOpts, llb.Readonly)
+		}
+
+		runOpts = append(runOpts, llb.AddMount(mnt.Target, srcSt, mountOpts...))
+	}
+
+	if opts.InsecureRootCapabilities {
+		runOpts = append(runOpts, llb.Security(llb.SecurityModeInsecure))
+	}
+
+	fsSt, err := container.FSState()
+	if err != nil {
+		return nil, fmt.Errorf("fs state: %w", err)
+	}
+
+	execSt := fsSt.Run(runOpts...)
+
+	execMD := buildkit.ExecutionMetadata{
+		SystemEnvNames: container.SystemEnvNames,
+		ClientID:       clientID,
+		OTELEnvs:       otelEnvs,
+	}
+	execMDOpt, err := execMD.AsConstraintsOpt()
+	if err != nil {
+		return nil, fmt.Errorf("execution metadata: %w", err)
+	}
+	marshalOpts := []llb.ConstraintsOpt{
+		llb.Platform(platform.Spec()),
+		execMDOpt,
+	}
+	execDef, err := execSt.Root().Marshal(ctx, marshalOpts...)
+	if err != nil {
+		return nil, fmt.Errorf("marshal root: %w", err)
+	}
+
+	container.FS = execDef.ToPB()
+
+	metaDef, err := execSt.GetMount(buildkit.MetaMountDestPath).Marshal(ctx, marshalOpts...)
+	if err != nil {
+		return nil, fmt.Errorf("get meta mount: %w", err)
+	}
+
+	container.Meta = metaDef.ToPB()
+
+	for i, mnt := range mounts {
+		if mnt.Tmpfs || mnt.CacheVolumeID != "" {
+			continue
+		}
+
+		mountSt := execSt.GetMount(mnt.Target)
+
+		// propagate any changes to regular mounts to subsequent containers
+		execMountDef, err := mountSt.Marshal(ctx, marshalOpts...)
+		if err != nil {
+			return nil, fmt.Errorf("propagate %s: %w", mnt.Target, err)
+		}
+
+		mounts[i].Source = execMountDef.ToPB()
+	}
+
+	container.Mounts = mounts
+
+	// set image ref to empty string
+	container.ImageRef = ""
+
+	return container, nil
+}
